# docs/swagger.yaml
basePath: /api
definitions:
...
  server_internal_controller_http_admin_request.CreateLink:
    properties:
      description:
        maxLength: 500
        type: string
      logo:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        maximum: 9999
        minimum: 0
        type: integer
      status:
        enum:
        - active
        - inactive
        type: string
      url:
        type: string
    required:
    - name
    - status
    - url
    type: object
...
  server_internal_controller_http_admin_request.GenerateUploadURL:
    properties:
      content_type:
        enum:
        - image/jpeg
        - image/png
        - image/webp
        type: string
      expiry_seconds:
        maximum: 86400
        minimum: 1
        type: integer
      file_name:
        maxLength: 255
        minLength: 1
        type: string
      file_size:
        maximum: 10485760
        minimum: 1
        type: integer
      resource_id:
        minimum: 0
        type: integer
      upload_type:
        enum:
        - avatar
        - post_cover
        - post_content
        type: string
    required:
    - content_type
    - file_name
    - file_size
    - upload_type
    type: object
...
  server_internal_controller_http_admin_request.UpdateLink:
    properties:
      description:
        maxLength: 500
        type: string
      id:
        minimum: 1
        type: integer
      logo:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        maximum: 9999
        minimum: 0
        type: integer
      status:
        enum:
        - active
        - inactive
        type: string
      url:
        type: string
    required:
    - id
    - name
    - status
    - url
    type: object
...
  server_internal_controller_http_admin_response.CreateLink:
    properties:
      id:
        type: integer
    type: object
...
  server_internal_controller_http_admin_response.FileDetail:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      object_key:
        type: string
      resource_id:
        type: integer
      uploader_id:
        type: integer
      url:
        type: string
      usage:
        type: string
    type: object
  server_internal_controller_http_admin_response.FileDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/server_internal_controller_http_admin_response.FileDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  server_internal_controller_http_admin_response.FileUploadURL:
    properties:
      expires:
        type: integer
      file_id:
        type: integer
      object_key:
        type: string
      upload_url:
        type: string
    type: object
...
  server_internal_controller_http_admin_response.LinkDetail:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  server_internal_controller_http_admin_response.LinkDetailPage:
    properties:
      current_page:
        type: integer
      list:
        items:
          $ref: '#/definitions/server_internal_controller_http_admin_response.LinkDetail'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  server_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: Nlxiaoy Blog 后端 API（Admin + Public V1）。
  title: Nlxiaoy Blog API
  version: "1.0"
paths:
...
  /admin/files/:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        enum:
        - created_at
        - file_size
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: 用途过滤
        in: query
        name: filter.usage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_admin_response.FileDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 文件列表
      tags:
      - Admin.Files
  /admin/files/{object_key}:
    delete:
      parameters:
      - description: 对象 Key
        in: path
        name: object_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 删除文件
      tags:
      - Admin.Files
  /admin/files/upload-url:
    post:
      consumes:
      - application/json
      parameters:
      - description: 上传参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_admin_request.GenerateUploadURL'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_admin_response.FileUploadURL'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 生成文件上传 URL
      tags:
      - Admin.Files
  /admin/links/:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 排序字段
        enum:
        - sort_order
        - created_at
        - name
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序方向
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_admin_response.LinkDetailPage'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 友链列表
      tags:
      - Admin.Links
    post:
      consumes:
      - application/json
      parameters:
      - description: 友链信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_admin_request.CreateLink'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_admin_response.CreateLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 创建友链
      tags:
      - Admin.Links
  /admin/links/{id}:
    delete:
      parameters:
      - description: 友链 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 删除友链
      tags:
      - Admin.Links
    put:
      consumes:
      - application/json
      parameters:
      - description: 友链 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 友链信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_admin_request.UpdateLink'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - AdminSession: []
      summary: 更新友链
      tags:
      - Admin.Links
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"